Move shared styles into 'utilities/styles'
